Adding new component to make search work on re-render
